Fix userAgent
